+++ resources/views/sales/read.blade.php
@@ -4,7 +4,6 @@
             <th>Kode Transaksi</th>
             <th>Tanggal</th>
             <th>Nama Sales</th>
-            <th>Total</th>
             <th>Aksi</th>
         </tr>
     </thead>
@@ -14,7 +13,6 @@
                 <td>{{ $sale->number }}</td>
                 <td>{{ $sale->date }}</td>
                 <td>{{ $sale->user->name }}</td>
-                <td>{{ $sale->date }}</td>
                 <td>
                     <div class="dropdown">
                         <button class="btn btn-outline-secondary btn-sm dropdown-toggle" type="button"
@@ -22,8 +20,12 @@
                         <ul class="dropdown-menu">
                             <li><a class="dropdown-item" onclick="showDetails({{ $sale->id }})">Details</a>
                             </li>
+                        @if ($user = Auth::user())
+                            @if ($user->role == 'superadmin' || $user->role == 'sales')
                             <li><a class="dropdown-item" onclick="show({{ $sale->id }})">Edit</a></li>
                             <li><a class="dropdown-item" onclick="deleteSales({{ $sale->id }})">Hapus</a></li>
+                            @endif
+                        @endif
                         </ul>
                     </div>
                 </td>
